# api-docs.yaml
openapi: 3.0.3
info:
  title: Geocoding API
  description: |
    A fast and comprehensive REST API for US ZIP code geocoding, Ohio address lookup, and county boundary data.
    
    ## 🔐 Authentication Required
    
    Most endpoints require an API key. To get started:
    
    1. **Register**: Visit the web interface and create an account
    2. **Get API Key**: Create an API key in your dashboard
    3. **Authenticate**: Use either authentication method:
       - `Authorization: Bearer your-api-key` header, OR
       - `X-API-Key: your-api-key` header
    4. **Test**: Use the "Authorize" button below to set your API key for all requests
    
    ## ✨ Features
    - **Lightning Fast**: PostgreSQL + PostGIS backed with optimized spatial indexes
    - **Multi-Level Coverage**: ZIP codes, cities, addresses, and county boundaries
    - **Comprehensive Data**: Population, density, coordinates, and geographic boundaries
    - **Flexible Search**: Multiple search modes with proximity and filtering
    - **ZIP Code Fallback**: City-based ZIP code lookup for flexible address searching
    - **Rich Metadata**: Timezone, county weights, administrative data
    - **GeoJSON Support**: Ready-to-use geographic data for mapping applications
    - **Usage Tracking**: Monitor your API usage and billing
    
    ## 📊 Data Coverage
    
    ### **🏙️ City Data**
    - **31,000+ US Cities**: Nationwide city coverage with demographics
    - **ZIP Code Associations**: All ZIP codes for each city
...
servers:
...
paths:
...
  /search:
    get:
      summary: Search ZIP Codes by City
      description: |
        Search for ZIP codes by city name with optional state filtering.
        
        Supports partial city name matching and returns multiple results.
        Useful for autocomplete functionality or when exact ZIP code is unknown.
      operationId: searchZipCodes
      security:
        - ApiKeyAuth: []
      tags:
        - Search
      parameters:
        - name: city
          in: query
          required: true
          description: City name or partial city name to search for
          schema:
            type: string
            minLength: 2
            example: "Springfield"
        - name: state
          in: query
          required: false
          description: Two-letter state code to filter results
          schema:
            type: string
            pattern: '^[A-Z]{2}$'
            example: "IL"
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
            example: 25
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              examples:
                springfield_search:
                  summary: Springfield search results
                  value:
                    success: true
                    data:
                      - zip_code: "62701"
                        city_name: "Springfield"
                        state_code: "IL"
                        state_name: "Illinois"
                        latitude: 39.78172
                        longitude: -89.64371
                      - zip_code: "62702"
                        city_name: "Springfield"
                        state_code: "IL"
                        state_name: "Illinois"
                        latitude: 39.76009
                        longitude: -89.60733
                    count: 2
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /addresses:
    get:
      summary: Search Ohio Addresses
      description: |
        Search for addresses within Ohio with flexible filtering options.
        
        Supports searching by address components, city, county, or proximity to coordinates.
        Perfect for address validation, autocomplete, and location-based services.
      operationId: searchOhioAddresses
      security:
        - ApiKeyAuth: []
      tags:
        - Ohio Addresses
      parameters:
        - name: query
          in: query
          required: false
          description: General search query (searches across address fields)
          schema:
            type: string
            example: "123 Main St"
        - name: house_number
          in: query
          required: false
          description: House number to search for
          schema:
            type: string
            example: "123"
        - name: street
          in: query
          required: false
          description: Street name to search for
          schema:
            type: string
            example: "Main St"
        - name: city
          in: query
          required: false
          description: City name to filter by
          schema:
            type: string
            example: "Columbus"
        - name: county
          in: query
          required: false
          description: County name to filter by
          schema:
            type: string
            example: "Franklin"
        - name: postcode
          in: query
          required: false
          description: ZIP code to filter by
          schema:
            type: string
            example: "43215"
        - name: lat
          in: query
          required: false
          description: Latitude for proximity search (requires lon and radius)
          schema:
            type: number
            format: double
            example: 39.9612
        - name: lon
          in: query
          required: false
          description: Longitude for proximity search (requires lat and radius)
          schema:
            type: number
            format: double
            example: -82.9988
        - name: radius
          in: query
          required: false
          description: Search radius in miles for proximity search
          schema:
            type: number
            format: double
            minimum: 0.1
            maximum: 50
            default: 1
            example: 5
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
            example: 50
        - name: offset
          in: query
          required: false
          description: Number of results to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
        - name: sort
          in: query
          required: false
          description: Sort order for results
          schema:
            type: string
            enum: [distance, address, city, county]
            default: address
            example: distance
      responses:
        '200':
          description: Address search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressSearchResponse'
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /counties:
    get:
      summary: List Ohio Counties
      description: |
        Retrieve a list of all Ohio counties with basic information and address counts.
        
        Supports filtering by name, address count ranges, and pagination.
        Perfect for building county selectors and administrative dashboards.
      operationId: getCounties
      security:
        - ApiKeyAuth: []
      tags:
        - County Boundaries
      parameters:
        - name: name
          in: query
          required: false
          description: Filter counties by name (partial match)
          schema:
            type: string
            example: "Franklin"
        - name: min_addresses
          in: query
          required: false
          description: Minimum number of addresses per county
          schema:
            type: integer
            minimum: 0
            example: 10000
        - name: max_addresses
          in: query
          required: false
          description: Maximum number of addresses per county
          schema:
            type: integer
            minimum: 0
            example: 500000
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
            example: 25
        - name: offset
          in: query
          required: false
          description: Number of results to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        '200':
          description: Counties retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountyListResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /cities:
    get:
      summary: Search US Cities
      description: |
        Search for US cities using various filters including name, state, county, population, and location-based queries.
        
        The endpoint supports fuzzy text matching on city names and can filter by:
        - City name (partial match)
        - State (by ID or name)
        - County name
        - Minimum population
        - Geographic proximity (latitude/longitude with radius)
        
        Results are ordered by ranking (major cities first) and population.
        
        **Use Case**: Use this endpoint as a fallback when ZIP code searches fail. For example,
        if searching for an address with an incorrect ZIP code, search for the city to get all
        valid ZIP codes for that city, then search addresses in each ZIP.
      operationId: searchCities
      security:
        - ApiKeyAuth: []
      tags:
        - Cities
      parameters:
        - name: query
          in: query
          required: false
          description: Free text search query for city name
          schema:
            type: string
            example: "Dayton"
        - name: city
          in: query
          required: false
          description: Filter by city name (supports fuzzy matching)
          schema:
            type: string
            example: "Dayton"
        - name: state
          in: query
          required: false
          description: Filter by state ID (e.g., "OH") or state name (e.g., "Ohio")
          schema:
            type: string
            example: "OH"
        - name: county
          in: query
          required: false
          description: Filter by county name
          schema:
            type: string
            example: "Montgomery"
        - name: lat
          in: query
          required: false
          description: Latitude for location-based search (requires lng and optionally radius)
          schema:
            type: number
            format: double
            example: 39.7589
        - name: lng
          in: query
          required: false
          description: Longitude for location-based search (requires lat and optionally radius)
          schema:
            type: number
            format: double
            example: -84.1916
        - name: radius
          in: query
          required: false
          description: Search radius in kilometers (requires lat and lng)
          schema:
            type: number
            format: double
            example: 50
        - name: min_population
          in: query
          required: false
          description: Minimum population filter
          schema:
            type: integer
            example: 100000
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return (max 100)
          schema:
            type: integer
            default: 10
            maximum: 100
            example: 10
        - name: offset
          in: query
          required: false
          description: Number of results to skip for pagination
          schema:
            type: integer
            default: 0
            example: 0
      responses:
        '200':
          description: Cities found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitySearchResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/counties:
    get:
      summary: Get County Statistics
      description: |
        **Admin endpoint** to retrieve comprehensive statistics about Ohio counties.
        
        Returns aggregated data including total counties, address counts, and distribution metrics.
      operationId: getCountyStats
      tags:
        - Admin
      responses:
        '200':
          description: County statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountyStatsResponse'
        '500':
          description: Failed to retrieve statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Supports two methods:
        
        **Method 1 - X-API-Key Header (Recommended):**
        ```
        X-API-Key: your-api-key-here
        ```
        
        **Method 2 - Authorization Bearer:**
        ```
        Authorization: Bearer your-api-key-here
        ```
        
        Get your API key by:
        1. Register/login at the web interface
        2. Create an API key in your dashboard
        3. Use either authentication method above
  schemas:
    ZipCode:
      type: object
      description: Complete ZIP code information including geographic and demographic data
      properties:
        zip_code:
          type: string
          description: The ZIP code identifier
          example: "10001"
        city_name:
          type: string
          description: Official USPS city name
          example: "New York"
        state_code:
          type: string
          description: Two-letter state abbreviation
          example: "NY"
        state_name:
          type: string
          description: Full state name
          example: "New York"
        zcta:
          type: boolean
          description: Whether this is a ZIP Code Tabulation Area
          example: true
        zcta_parent:
          type: string
          nullable: true
          description: Parent ZCTA if applicable
          example: null
        population:
          type: number
          nullable: true
          description: Population count for this ZIP code area
          example: 21102.0
        density:
          type: number
          nullable: true
          description: Population density per square mile
          example: 35400.5
        primary_county_code:
          type: string
          description: FIPS code of the primary county
          example: "36061"
        primary_county_name:
          type: string
          description: Name of the primary county
          example: "New York"
        county_weights:
          type: object
          description: JSON object with county codes and their percentage weights
          additionalProperties:
            type: string
          example:
            "36061": "100"
        county_names:
          type: array
          items:
            type: string
          description: Array of all county names for this ZIP code
          example: ["New York"]
        county_codes:
          type: array
          items:
            type: string
          description: Array of all county FIPS codes for this ZIP code
          example: ["36061"]
        imprecise:
          type: boolean
          description: Whether the coordinates are imprecise
          example: false
        military:
          type: boolean
          description: Whether this is a military ZIP code
          example: false
        timezone:
          type: string
          description: IANA timezone identifier
          example: "America/New_York"
        latitude:
          type: number
          format: double
          description: Latitude coordinate (WGS84)
          example: 40.75066
        longitude:
          type: number
          format: double
          description: Longitude coordinate (WGS84)
          example: -73.99670
...
    SearchResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ZipCode'
          description: Array of matching ZIP codes
        count:
          type: integer
          description: Number of results returned
          example: 2
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
          example: false
        error:
          type: string
          description: Error message describing what went wrong
          example: "ZIP code not found"
...
    OhioAddress:
      type: object
      description: Complete Ohio address information with geographic coordinates
      properties:
        id:
          type: integer
          description: Unique address identifier
          example: 12345
        hash:
          type: string
          description: Unique hash for the address
          example: "abc123def456"
        house_number:
          type: string
          description: House number
          example: "123"
        street:
          type: string
          description: Street name
          example: "Main St"
        unit:
          type: string
          nullable: true
          description: Unit or apartment number
          example: "Apt 2B"
        city:
          type: string
          description: City name
          example: "Columbus"
        district:
          type: string
          nullable: true
          description: District or neighborhood
          example: "Downtown"
        region:
          type: string
          description: State or region
          example: "Ohio"
        postcode:
          type: string
          description: ZIP code
          example: "43215"
        county:
          type: string
          description: County name
          example: "Franklin"
        latitude:
          type: number
          format: double
          description: Latitude coordinate (WGS84)
          example: 39.9612
        longitude:
          type: number
          format: double
          description: Longitude coordinate (WGS84)
          example: -82.9988
    AddressSearchResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/OhioAddress'
              - type: object
                properties:
                  distance_miles:
                    type: number
                    format: double
                    description: Distance in miles (only present for proximity searches)
                    example: 0.5
        count:
          type: integer
          description: Number of addresses returned
          example: 25
        total:
          type: integer
          description: Total number of matching addresses
          example: 150
...
    CountyBasic:
      type: object
      description: Basic county information
      properties:
        id:
          type: integer
          description: Unique county identifier
          example: 1
        county_name:
          type: string
          description: County name
          example: "Franklin"
        address_count:
          type: integer
          description: Number of addresses in this county
          example: 852417
...
    CountyListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/CountyBasic'
        count:
          type: integer
          description: Number of counties returned
          example: 25
...
    CountyStatsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            total_counties:
              type: integer
              description: Total number of Ohio counties
              example: 89
            total_addresses:
              type: integer
              description: Total number of addresses across all counties
              example: 5551283
            avg_addresses_per_county:
              type: number
              format: double
              description: Average addresses per county
              example: 62373.6
            max_addresses_per_county:
              type: integer
              description: Maximum addresses in a single county
              example: 852417
            min_addresses_per_county:
              type: integer
              description: Minimum addresses in a single county
              example: 1832
...
    City:
      type: object
      description: US city information including location, demographics, and associated ZIP codes
      properties:
        id:
          type: integer
          format: int64
          description: Unique city identifier
          example: 12345
        city:
          type: string
          description: City name
          example: "Dayton"
        city_ascii:
          type: string
          description: ASCII-only version of city name
          example: "Dayton"
        state_id:
          type: string
          description: Two-letter state code
          example: "OH"
        state_name:
          type: string
          description: Full state name
          example: "Ohio"
        county_fips:
          type: string
          description: County FIPS code
          example: "39113"
        county_name:
          type: string
          description: County name
          example: "Montgomery"
        lat:
          type: number
          format: double
          description: Latitude coordinate
          example: 39.7589
        lng:
          type: number
          format: double
          description: Longitude coordinate
          example: -84.1916
        population:
          type: integer
          description: City population
          example: 140407
        density:
          type: number
          format: double
          description: Population density (people per sq km)
          example: 1234.56
        source:
          type: string
          description: Data source
          example: "shape"
        military:
          type: boolean
          description: Whether this is a military base
          example: false
        incorporated:
          type: boolean
          description: Whether the city is incorporated
          example: true
        timezone:
          type: string
          description: IANA timezone identifier
          example: "America/New_York"
        ranking:
          type: integer
          description: City ranking (1 = major city)
          example: 1
        zips:
          type: string
          description: Space-separated list of ZIP codes
          example: "45401 45402 45403 45404 45405 45406 45409 45410 45414 45415"
        external_id:
          type: string
          description: External identifier from source data
          example: "1840034016"
    CitySearchResponse:
      type: object
      description: Response for city search operations
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/City'
        count:
          type: integer
          description: Number of results in current response
          example: 1
        total:
          type: integer
          description: Total number of matching cities
          example: 1
        query:
          type: string
          description: Original search query
          example: "Dayton"
        filters:
          type: object
          description: Applied filters
          additionalProperties: true
          example:
            state: "OH"
        message:
          type: string
          description: Optional message
        error:
          type: string
          description: Error message if success is false
...